backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.1
python-multipart>=0.0.9
python-dotenv>=1.0.1
httpx>=0.26.0
pytest>=8.0.0
sqlalchemy>=2.0.0
# AI Libraries (uncomment as needed)
# openai>=1.12.0
groq>=0.4.0
moviepy>=1.0.3
python-jose[cryptography]>=3.3.0
uuid>=1.30
# For production
# gunicorn>=21.2.0 